Reuse top crop to show previews the same as before
